task4-llm-extraction/requirements.txt
openai>=1.54.0
python-dotenv>=1.0.0
pandas>=2.0.0
python-dateutil>=2.8.0
pytest>=7.0.0
